requirements.txt

Django>=6.0.2
pillow
jazzmin
django-import-export
